Fix doc link and tests
