+++ src/platform/packages/shared/kbn-doc-links/src/get_doc_links.ts
@@ -678,7 +678,7 @@ export const getDocLinks = ({ kibanaBranch, buildFlavor }: GetDocLinkOptions): D
         ? `${SERVERLESS_DOCS}observability-create-slo-burn-rate-alert-rule.html`
         : `${OBSERVABILITY_DOCS}slo-burn-rate-alert.html`,
       aiAssistant: `${OBSERVABILITY_DOCS}obs-ai-assistant.html`,
-      elasticManagedLlm: `${DOCS_WEBSITE_URL}reference/kibana/connectors-kibana/elastic-managed-llm`,
+      elasticManagedLlm: `${ELASTIC_WEBSITE_URL}docs/reference/kibana/connectors-kibana/elastic-managed-llm`,
       elasticManagedLlmUsageCost: `${ELASTIC_WEBSITE_URL}pricing`,
     },
     alerting: {
